Refactor legacy greeting module layout
