+++ crates/protocol/src/legacy/greeting/mod.rs
@@ -1,0 +1,15 @@
+mod format;
+mod parse;
+mod tokens;
+mod types;
+
+pub use format::{format_legacy_daemon_greeting, write_legacy_daemon_greeting};
+pub use parse::{
+    parse_legacy_daemon_greeting, parse_legacy_daemon_greeting_details,
+    parse_legacy_daemon_greeting_owned,
+};
+pub use tokens::DigestListTokens;
+pub use types::{LegacyDaemonGreeting, LegacyDaemonGreetingOwned};
+
+#[cfg(test)]
+mod tests;
